# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "Adjustment": {
        "enum": [
          "splits",
          "none",
          "total_return"
        ],
        "title": "Adjustment",
        "type": "string"
      },
...
      "Bar": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "session_date": {
            "format": "date",
            "title": "Session Date",
            "type": "string"
          },
          "ts": {
            "title": "Ts",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "number"
          },
          "vwap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vwap"
          }
        },
        "required": [
          "ts",
          "session_date",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "Bar",
        "type": "object"
      },
      "BarSeries": {
        "properties": {
          "adjustment": {
            "title": "Adjustment",
            "type": "string"
          },
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Bar"
            },
            "title": "Items",
            "type": "array"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "adjustment",
          "start",
          "end",
          "items"
        ],
        "title": "BarSeries",
        "type": "object"
      },
...
      "FeatureSeries": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "title": "Names",
            "type": "array"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "names",
          "start",
          "end",
          "items"
        ],
        "title": "FeatureSeries",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstrumentDetail": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "feature_sessions": {
            "additionalProperties": {
              "format": "date",
              "type": "string"
            },
            "title": "Feature Sessions",
            "type": "object"
          },
          "features": {
            "additionalProperties": true,
            "title": "Features",
            "type": "object"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "reference": {
            "additionalProperties": true,
            "title": "Reference",
            "type": "object"
          },
          "reference_snapshot": {
            "format": "date",
            "title": "Reference Snapshot",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "reference_snapshot",
          "reference",
          "company",
          "features",
          "feature_sessions"
        ],
        "title": "InstrumentDetail",
        "type": "object"
      },
      "InstrumentEvent": {
        "properties": {
          "table": {
            "title": "Table",
            "type": "string"
          },
          "ts": {
            "title": "Ts",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "table",
          "ts",
          "values"
        ],
        "title": "InstrumentEvent",
        "type": "object"
      },
...
      "OptionChain": {
        "properties": {
          "expiries": {
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Expiries",
            "type": "array"
          },
          "our_iv": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "our iv30 rollup row for the session (decimal)",
            "title": "Our Iv"
          },
          "quotes": {
            "description": "sorted by expiry, strike, right",
            "items": {
              "$ref": "#/components/schemas/OptionQuote"
            },
            "title": "Quotes",
            "type": "array"
          },
          "session": {
            "format": "date",
            "title": "Session",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "the chain fetch status (OK, NO_CHAIN, STALE_DATA: ...)",
            "title": "Status"
          },
          "strikes": {
            "items": {
              "type": "number"
            },
            "title": "Strikes",
            "type": "array"
          },
          "underlying": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "the underlying quote with the chain (Cboe iv30 in percent)",
            "title": "Underlying"
          },
          "underlying_id": {
            "title": "Underlying Id",
            "type": "string"
          }
        },
        "required": [
          "underlying_id",
          "session",
          "status",
          "underlying",
          "our_iv",
          "expiries",
          "strikes",
          "quotes"
        ],
        "title": "OptionChain",
        "type": "object"
      },
      "OptionQuote": {
        "properties": {
          "ask": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask"
          },
          "bid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "expiry": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry"
          },
          "gamma": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gamma"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "iv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv"
          },
          "last": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last"
          },
          "open_interest": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Interest"
          },
          "rho": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rho"
          },
          "right": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Right"
          },
          "strike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike"
          },
          "theta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theta"
          },
          "vega": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vega"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "instrument_id",
          "expiry",
          "right",
          "strike",
          "bid",
          "ask",
          "last",
          "volume",
          "open_interest",
          "iv",
          "delta",
          "gamma",
          "theta",
          "vega",
          "rho"
        ],
        "title": "OptionQuote",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only API over the algotrade stores (ADR 0025).",
    "title": "algotrade API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chains/{underlying_id}": {
      "get": {
        "operationId": "chain_chains__underlying_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "underlying_id",
            "required": true,
            "schema": {
              "title": "Underlying Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "expiry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expiry"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionChain"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chain",
        "tags": [
          "chains"
        ]
      }
    },
...
    "/instruments/{instrument_id}": {
      "get": {
        "operationId": "detail_instruments__instrument_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "instruments"
        ]
      }
    },
    "/instruments/{instrument_id}/bars": {
      "get": {
        "operationId": "bars_instruments__instrument_id__bars_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "adjust",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Adjustment",
              "default": "splits"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarSeries"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bars",
        "tags": [
          "instruments"
        ]
      }
    },
    "/instruments/{instrument_id}/events": {
      "get": {
        "operationId": "events_instruments__instrument_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstrumentEvent"
                  },
                  "title": "Response Events Instruments  Instrument Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "instruments"
        ]
      }
    },
    "/instruments/{instrument_id}/features": {
      "get": {
        "operationId": "features_instruments__instrument_id__features_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "description": "comma-separated field names",
            "in": "query",
            "name": "names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "comma-separated field names",
              "title": "Names"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureSeries"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Features",
        "tags": [
          "instruments"
        ]
      }
    },
...
  }
}